+++ cmd/cronosd/cmd/root.go
@@ -12,6 +12,7 @@ import (
 	rosettaCmd "github.com/cosmos/rosetta/cmd"
 	memiavlcfg "github.com/crypto-org-chain/cronos/store/config"
 	"github.com/crypto-org-chain/cronos/v2/app"
+	config2 "github.com/crypto-org-chain/cronos/v2/cmd/cronosd/config"
 	"github.com/crypto-org-chain/cronos/v2/cmd/cronosd/opendb"
 	"github.com/crypto-org-chain/cronos/v2/x/cronos"
 	e2eecli "github.com/crypto-org-chain/cronos/v2/x/e2ee/client/cli"
@@ -272,6 +273,7 @@ func initAppConfig() (string, interface{}) {
 
 		MemIAVL   memiavlcfg.MemIAVLConfig `mapstructure:"memiavl"`
 		VersionDB VersionDBConfig          `mapstructure:"versiondb"`
+		Cronos    config2.CronosConfig     `mapstructure:"cronos"`
 	}
 
 	tpl, cfg := servercfg.AppConfig("")
@@ -280,9 +282,10 @@ func initAppConfig() (string, interface{}) {
 		Config:    cfg.(servercfg.Config),
 		MemIAVL:   memiavlcfg.DefaultMemIAVLConfig(),
 		VersionDB: DefaultVersionDBConfig(),
+		Cronos:    config2.DefaultCronosConfig(),
 	}
 
-	return tpl + memiavlcfg.DefaultConfigTemplate + DefaultVersionDBTemplate, customAppConfig
+	return tpl + memiavlcfg.DefaultConfigTemplate + DefaultVersionDBTemplate + config2.DefaultCronosConfigTemplate, customAppConfig
 }
 
 // newApp creates the application