+++ cmd/cronosd/config/config.go
@@ -8,3 +8,17 @@ func SetBech32Prefixes(config *sdk.Config) {
 	config.SetBech32PrefixForValidator(Bech32PrefixValAddr, Bech32PrefixValPub)
 	config.SetBech32PrefixForConsensusNode(Bech32PrefixConsAddr, Bech32PrefixConsPub)
 }
+
+type CronosConfig struct {
+	// Set to true to disable tx replacement.
+	DisableTxReplacement bool `mapstructure:"disable-tx-replacement"`
+	// Set to true to disable optimistic execution.
+	DisableOptimisticExecution bool `mapstructure:"disable-optimistic-execution"`
+}
+
+func DefaultCronosConfig() CronosConfig {
+	return CronosConfig{
+		DisableTxReplacement:       false,
+		DisableOptimisticExecution: false,
+	}
+}